addressing some PR feedback
