+++ test/NuGet.Clients.Tests/NuGet.CommandLine.Test/RestoreTransitiveLoggingTests.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
@@ -668,7 +668,7 @@ namespace NuGet.CommandLine.Test
 
                 // Assert
                 r.Success.Should().BeTrue();
-                r.AllOutput.GetSubstringCount("NU1603", ignoreCase: false).Should().Be(3);
+                GetSubstringCount(r.AllOutput, "NU1603", ignoreCase: false).Should().Be(3);
             }
         }
 
@@ -907,7 +907,7 @@ namespace NuGet.CommandLine.Test
 
                 // Assert
                 r.Success.Should().BeTrue();
-                r.AllOutput.GetSubstringCount("NU1603", ignoreCase: false).Should().Be(1);
+                GetSubstringCount(r.AllOutput, "NU1603", ignoreCase: false).Should().Be(1);
             }
         }
 
@@ -988,7 +988,7 @@ namespace NuGet.CommandLine.Test
 
                 // Assert
                 r.Success.Should().BeTrue();
-                r.AllOutput.GetSubstringCount("NU1603", ignoreCase: false).Should().Be(2);
+                GetSubstringCount(r.AllOutput, "NU1603", ignoreCase: false).Should().Be(2);
             }
         }
 
@@ -1174,6 +1174,19 @@ namespace NuGet.CommandLine.Test
                 r.Success.Should().BeTrue();
                 r.AllOutput.Should().Contain("NU1603");
             }
+        }
+
+        private static int GetSubstringCount(string str, string substr, bool ignoreCase)
+        {
+            var splitChars = new char[] { '.', '?', '!', ' ', ';', ':', ',', '\r', '\n' };
+            var words = str.Split(splitChars, StringSplitOptions.RemoveEmptyEntries);
+            var comparisonType = ignoreCase ?
+                StringComparison.OrdinalIgnoreCase :
+                StringComparison.Ordinal;
+
+            return words
+                .Where(word => string.Equals(word, substr, comparisonType))
+                .Count();
         }
     }
 }
